+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/button/ControlHandler.java
@@ -35,7 +35,7 @@ abstract class ControlHandler implements Consumer<Action> {
 			onSeparator();
 		}
 		else if (action instanceof Controls) {
-			if (((Controls) action).notEmpty()) {
+			if (((Controls) action).size() > 0) {
 				onControls((Controls) action);
 			}
 		}