+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/button/DefaultMenuBuilder.java
@@ -155,7 +155,7 @@ final class DefaultMenuBuilder extends AbstractComponentBuilder<Void, JMenu, Men
 
 		@Override
 		public boolean test(Controls controls) {
-			return controls.notEmpty();
+			return controls.size() > 0;
 		}
 	}